supporting resource limitation and monitoring in job and supporting fmt on cpu core calculation with scalability and proper synchronization with parallel processes integrated with proper crash hanndling and also adding flexibility on job operatons as well as job utilities

+++ exceptions/resources/cpu_violations.h
@@ -1,6 +1,6 @@
 #include "../../base/config.h"
 #define MAX_CPU_CORE 10
-__attribute__((hot)) float convertToFloat(char* cpuLimit, int* err_flg);
+__attribute__((hot)) double convertToFloat(char* cpuLimit, int* err_flg);
