+++ res_man/level_utility/extreme.c
@@ -23,7 +23,7 @@ DECLARE_128_T
 CREATE_PROCESS_REG_F
 CREATE_PROCESS__F_V1
 
-__attribute__((hot)) int limitCpuAndMemory_F_EXTR(char* serviceName, __uint32_t servicePid, float cpuLimit, _Bool limitMemory, __uint128_t memBytes) {
+__attribute__((hot)) int limitCpuAndMemory_F_EXTR(char* serviceName, __uint32_t servicePid, double cpuLimit, _Bool limitMemory, __uint128_t memBytes) {
     if (access("/sys/fs/cgroup/cgroup.controllers", F_OK) == 0) {
         if(__builtin_expect(basic_v2_setup(serviceName) != 0, 0)) {
             return -1;
