+++ res_man/level_utility/low.h
@@ -1,2 +1,2 @@
 #include <stdio.h>
-__attribute__((hot)) int limitCpuAndMemory(char* serviceName, __uint32_t servicePid, float cpuLimit, _Bool limitMemory, int memBytes);
+__attribute__((hot)) int limitCpuAndMemory(char* serviceName, __uint32_t servicePid, double cpuLimit, _Bool limitMemory, int memBytes);
